parser/requirements.txt
Scrapy==2.12.0
lxml==5.3.0
cssselect==1.2.0
parsel==1.10.0
itemadapter==0.10.0
itemloaders==1.3.2
w3lib==2.2.1
twisted==24.11.0
queuelib==1.7.0
protego==0.4.0
tldextract==5.1.3
httpx==0.28.1
aiosqlite==0.20.0
pydantic==2.10.4
python-dotenv==1.0.1
sqlalchemy==2.0.36
asyncpg==0.30.0
inflection==0.5.1
fastapi-users-db-sqlalchemy==7.0.0